test: use fail-rs for some tests. (#2592)
